Add make static functionality to inner fields
Adds functionality to the API for editable fields within static math
fields (inner math fields) to make them static or editable.
makeStatic() changes the class of the editable field to
“mq-editable-field-static”, removes its texture, and unbinds its event
handlers. makeEditable() does the opposite. It wraps the functions for
initializing an editable field in an easy-to-reference way.

This approach maintains the existing tree structure and any references
to the inner editable field in the static math wrapper’s inner fields
array. This approach assumes that a static math field will take over
the event handling for an editable field after it has been made static.